makefile: correctly sed paths
The default Makefile for tlp makes a mess with catenating `DESTDIR` to
everything, but then not actualy using the catenated (_ prefixed)
variables to sed it's `.in` files.

This patch makes sure that it correctly sets the paths, taking `DESTDIR`
in account where it makes sense (e.g. /bin where we want $out/bin) but
not where it doesn't (/etc/tlp.conf should be just that).

The reason DESTDIR is used at all, as opposed to the more appropriate
PREFIX, is covered in the nix formula, and is (also) due to the Makefile
being a bit "different."